refactor

+++ native/main/gfx/Capabilities.hpp
@@ -7,10 +7,10 @@ namespace gfx
     class Capabilities
     {
     public:
-        const uint32_t uniformBufferOffsetAlignment;
+        const size_t uniformBufferOffsetAlignment;
         const int clipSpaceMinZ;
 
-        Capabilities(uint32_t uniformBufferOffsetAlignment, int clipSpaceMinZ)
+        Capabilities(size_t uniformBufferOffsetAlignment, int clipSpaceMinZ)
             : uniformBufferOffsetAlignment(uniformBufferOffsetAlignment),
               clipSpaceMinZ(clipSpaceMinZ) {}
     };
